playground: refactor app into modules
Fixes #55.
