+++ examples/playground/src/playground/constants.ts
@@ -1,0 +1,8 @@
+export const ROOT_ID = "00000000000000000000000000000000"; // 16-byte zero, hex-encoded
+
+export const MAX_COMPOSER_NODE_COUNT = 100_000;
+
+export const PLAYGROUND_SYNC_MAX_CODEWORDS = 2_000_000;
+export const PLAYGROUND_SYNC_MAX_OPS_PER_BATCH = 20_000;
+export const PLAYGROUND_PEER_TIMEOUT_MS = 30_000;
+
